+++ src/components/Tables/Jobs/index.js
@@ -1,9 +1,7 @@
-export { default as JobsTabSwitcher } from './JobsTabSwitcher.react';
 export { default as getJobsColumns } from './getJobsColumns.react';
 
 export { default as NodeLogs } from './NodeLogs.react';
 export { default as NodeInputOutput } from './NodeInputOutput.react';
 export { default as NodeInfo } from './NodeInfo.react';
 
-export { default as JobGraph } from './JobGraph.react';
 export { default as JobInfo } from './JobInfo.react';
